Fix middleware type on EntryServer
