+++ packages/vite/src/middleware/types.ts
@@ -18,3 +18,8 @@ export type MiddlewareInvokeOptions = {
   cssPaths?: Array<string>
   params?: Record<string, unknown>
 }
+
+// Tuple of [mw, '*.{extension}']
+export type MiddlewareReg = Array<
+  [Middleware | MiddlewareClass, string] | Middleware | MiddlewareClass
+>
