+++ packages/vite/src/types.ts
@@ -10,7 +10,27 @@
  * **All** of these properties are used by the prod FE server
  */
 import type { RWRouteManifestItem } from '@redwoodjs/internal'
+import type { TagDescriptor } from '@redwoodjs/web'
+
+import type { MiddlewareReg } from './middleware/types'
 
 export type RWRouteManifest = Record<PathDefinition, RWRouteManifestItem>
 
 type PathDefinition = string
+
+export type ServerEntryType = React.FunctionComponent<{
+  css: string[]
+  meta: TagDescriptor[]
+}>
+
+export type EntryServer =
+  | {
+      registerMiddleware?: () => MiddlewareReg
+      ServerEntry: ServerEntryType
+      default: never
+    }
+  | {
+      registerMiddleware?: () => MiddlewareReg
+      ServerEntry: never
+      default: ServerEntryType
+    }
