+++ packages/vite/src/utils.ts
@@ -1,4 +1,8 @@
+import type { ViteDevServer } from 'vite'
+
 import { getPaths } from '@redwoodjs/project-config'
+
+import type { EntryServer } from './types'
 
 export function stripQueryStringAndHashFromPath(url: string) {
   return url.split('?')[0].split('#')[0]
@@ -19,7 +23,22 @@ export function ensureProcessDirWeb(webDir: string = getPaths().web.base) {
     process.chdir(webDir)
   }
 }
+
 export function makeFilePath(path: string): string {
   // Without this, absolute paths can't be imported on Windows
   return 'file:///' + path
+}
+
+export async function ssrLoadEntryServer(viteDevServer: ViteDevServer) {
+  const rwPaths = getPaths()
+
+  if (!rwPaths.web.entryServer) {
+    throw new Error('entryServer not defined')
+  }
+
+  return viteDevServer.ssrLoadModule(
+    rwPaths.web.entryServer,
+    // Have to type cast here because ssrLoadModule just returns a generic
+    // Record<string, any> type
+  ) as Promise<EntryServer>
 }
